% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_sju_mco.R
\docType{data}
\name{flights_sju_mco}
\alias{flights_sju_mco}
\title{Tidied Flights Data}
\format{A data frame
\describe{
    \item{seats}{Number of Seats}
    \item{passengers}{Number of passengers}
    \item{distance}{Distance in miles}
    \item{air_time}{Flight time (minutes)}
    \item{unique_carrier}{Abreviation of Carrier}
    \item{unique_carrier_name}{Full Name of Carrier}
    \item{origin}{Origin ["SJU" or "MCO"]}
    \item{dest}{Destination ["SJU" or "MCO"]}
    \item{aircraft_type}{Number of Aircraft accoridng to BTS}
    \item{year}{Year (1990-2019)}
    \item{month}{Month of the year (1-12)}
    }}
\usage{
flights_sju_mco
}
\description{
This data is an excerpt from the data published by the IEPR.
}
\keyword{datasets}
