% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightsPR.R
\docType{package}
\name{flightsPR}
\alias{flightsPR}
\alias{flightsPR-package}
\title{Clean up and share some data from flights from PR.}
\description{
Clean up and share some data from flights from PR.
}
\author{
Ian Flores Siaca, \email{iflores.siaca@gmail.com}
}
